+++ railties/CHANGELOG.md
@@ -1,3 +1,5 @@
+## Rails 4.0.6 (unreleased) ##
+
 *   Do not set the Rails environment to test by default when using test_unit Railtie.
 
     *Konstantin Shabanov*